Refactor: moved ip.py back to vm_supervisor and removed network package
